Fix testcase bugs found with choose satisfiability check

+++ frontends/scalac/src/it/resources/regression/verification/valid/Choose2.scala
@@ -1,13 +1,12 @@
 /* Copyright 2009-2016 EPFL, Lausanne */
 
 import stainless.lang._
-import stainless.lang.synthesis._
 
 object Choose2 {
 
-  def test(x: Int): Int = {
+  def test(x: BigInt): BigInt = {
 
-    choose((y: Int) => {
+    choose((y: BigInt) => {
       val z = y + 2
       z == y
     })
